Split up the `fxa-client/lib.rs` code
This is prep for the upcoming FxA client work.  I think the smaller
modules are easier to understand and will make it easier to communicate
changes.  Also, it's useful to know which structs go together with which
methods.
